LayoutTests/imported/w3c/web-platform-tests/mixed-content/gen/top.meta/opt-in/worklet-layout.https-expected.txt


FAIL Mixed-Content: Expects allowed for worklet-layout to same-https origin and no-redirect redirection from https context. assert_equals: The resource request should be 'allowed'. expected "allowed" but got "blocked"
PASS Mixed-Content: Expects blocked for worklet-layout to cross-http origin and no-redirect redirection from https context.
PASS Mixed-Content: Expects blocked for worklet-layout to same-http origin and no-redirect redirection from https context.
